+++ R/lbfgsb3c.R
@@ -102,7 +102,7 @@ lbfgsb3cControl <- function(trace=0,
                             literalFix=TRUE,
                             literalFixRes=TRUE,
                             addProp = c("combined2", "combined1"),
-                            calcTables=TRUE, compress=FALSE,
+                            calcTables=TRUE, compress=TRUE,
                             covMethod=c("r", ""),
                             adjObf=TRUE, ci=0.95, sigdig=4, sigdigTable=NULL, ...) {