+++ tests/testthat/test-augpred.R
@@ -1,9 +1,9 @@
 nmTest({
   test_that("test augPred", {
 
-    PKdata <- nlmixr2data::warfarin %>%
-      dplyr::filter(dvid == "cp") %>%
-      dplyr::select(-dvid) %>%
+    PKdata <- nlmixr2data::warfarin |>
+      dplyr::filter(dvid == "cp") |>
+      dplyr::select(-dvid) |>
       dplyr::mutate(sex = ifelse(sex == "male", 1, 0))
 
     One.comp.KA.solved <- function() {
@@ -48,7 +48,7 @@ nmTest({
           DUR = c(2.5, 2.5, NA, 2.5, 2.5, NA, 2.5, 2.5, NA, 2.5, 2.5, NA),
           DV = c(NA, NA, 3.0, NA, NA, 9.6, NA, NA, 7.0, NA, NA, 2.8),
           WT = c(rep(55, 6), rep(48, 6))
-        ) %>%
+        ) |>
         dplyr::mutate(EVID = ifelse(is.na(DV), 1, 0))
 
       fun <- function() {
@@ -76,21 +76,21 @@ nmTest({
 
   test_that("test augPred with xgxr dataset", {
 
-    dat <- xgxr::case1_pkpd %>%
-      dplyr::rename(DV=LIDV) %>%
-      dplyr::filter(CMT %in% 1:2) %>%
+    dat <- xgxr::case1_pkpd |>
+      dplyr::rename(DV=LIDV) |>
+      dplyr::filter(CMT %in% 1:2) |>
       dplyr::filter(TRTACT != "Placebo")
 
       doses <- unique(dat$DOSE)
       nid <- 3 # 7 ids per dose group
       dat2 <- do.call("rbind",
                       lapply(doses, function(x) {
-                        ids <- dat %>%
-                          dplyr::filter(DOSE == x) %>%
-                          dplyr::reframe(ids=unique(ID)) %>%
+                        ids <- dat |>
+                          dplyr::filter(DOSE == x) |>
+                          dplyr::reframe(ids=unique(ID)) |>
                           dplyr::pull()
                         ids <- ids[seq(1, nid)]
-                        dat %>%
+                        dat |>
                           dplyr::filter(ID %in% ids)
                       }))
 
@@ -221,7 +221,7 @@ nmTest({
       })
     }
 
-    model.1compt.depot1 <- model.1compt.depot %>%
+    model.1compt.depot1 <- model.1compt.depot |>
       ini(eta.ka~0)
 
     fit1 <- .nlmixr(model.1compt.depot1, theo_sd, est="focei",