+++ tests/testthat/test-cov-focei.R
@@ -3,7 +3,7 @@ nmTest({
   test_that("cov-focei", {
 
     dat <-
-      warfarin %>%
+      warfarin |>
       dplyr::filter(dvid == "cp")
 
       #### doesn't work with FOCEI ### doesn't work with SAEM with CRWES=True but does iwth CWRES=FALSE