+++ NEWS.md
@@ -1,3 +1,16 @@
+# nlmixr2est (development version)
+
+- Add `predict(fit, level="ipred")`, `predict(fit,
+  level="individual")` or `predict(fit, level=1)` to predict
+  individual fits (with possibly a new dataset).
+
+- Change test files to `.rds` files
+
+- Drop magrittr `%>%` in favor of `|>`.
+
+- Turn back on object compress and use roxde2's default for
+  compression.
+
 # nlmixr2est 5.0.0
 
 - Remove `qs` and change to `qs2`.  This breaks backward